new file:   Drivers/Basic/i686/debug.h x86debug工具 	new file:   Drivers/Basic/i686/io.c io.c 一些IO处理 	new file:   Drivers/Basic/i686/io.h io.c的头文件 	renamed:    types.h -> Drivers/Basic/i686/types.h 放在Drivers里了 	modified:   Makefile MAKEFILE通用适配 	modified:   kernel.c 把debug，io什么的放了DRIVERS
