+++ Drivers/Basic/i686/io.c
@@ -1,0 +1,20 @@
+//=============================================================================
+// ■ io.c
+//-----------------------------------------------------------------------------
+//   Driver > Basic > i686 > io
+//=============================================================================
+
+//---------------------------------------------------------------------------
+// ● 根据地址写入内存
+//---------------------------------------------------------------------------
+void write_memory(int address, char value) {
+	*((char*) address) = value;
+}
+
+//---------------------------------------------------------------------------
+// ● CPU 挂起
+//---------------------------------------------------------------------------
+void cpu_hlt () {
+	__asm__ ("hlt");
+}
+
