+++ Drivers/Basic/i686/io.h
@@ -1,0 +1,8 @@
+//=============================================================================
+// ■ io.h
+//-----------------------------------------------------------------------------
+//   Driver > Basic > i686 > io
+//=============================================================================
+
+void write_memory(int address, char value);
+void cpu_hlt();
